+++ src/renderer/Layout.tsx
@@ -108,12 +108,23 @@ const Layout: React.FC<IProps> = (props: IProps) => {
       <Box
         sx={{
           display: 'flex',
-          flexGrow: 1,
-          alignItems: 'center',
-          justifyContet: 'center',
           flexDirection: 'column',
-          width: '100%',
+          overflowY: 'scroll',
           height: '100%',
+          width: '100%',
+          scrollbarWidth: 'thin',
+          '&::-webkit-scrollbar': {
+            width: '1em',
+          },
+          '&::-webkit-scrollbar-track': {
+            background: '#f1f1f1',
+          },
+          '&::-webkit-scrollbar-thumb': {
+            backgroundColor: '#888',
+          },
+          '&::-webkit-scrollbar-thumb:hover': {
+            background: '#555',
+          },
         }}
       >
         <HomePage videoState={videoState} setNavigation={setNavigation} />