+++ docs/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Changed
 - [Issue 370](https://github.com/aza547/wow-recorder/issues/370) - Show more than the most recent video on home page.
+- [Issue 371](https://github.com/aza547/wow-recorder/issues/371) - Include Battleground names in the UI.
+- [Issue 352](https://github.com/aza547/wow-recorder/issues/352) - Change the home page to be less fancy and more functional.
+
 ### Fixed
 
 ## [3.10.1] - 2023-04-11